+++ tests/integration/simconnect_mock_test.go
@@ -147,6 +147,7 @@ func TestSimconnectToolsList(t *testing.T) {
 		"get_waypoints_in_range",
 		"get_waypoint_details",
 		"get_airport_taxiways",
+		"get_taxiway_names",
 		"get_airport_parkings",
 	}